Stretch 2 - Jeffrey Mejia Cortez

+++ server/server.ts
@@ -183,6 +183,22 @@ app.get('/api/carousel', async (req, res, next) => {
   }
 });
 
+app.delete('/api/checkout', authMiddleware, async (req, res, next) => {
+  try {
+    const sql = `
+      delete from "cartItems"
+      where "userId" = $1
+      returning*
+    `;
+    const results = await db.query(sql, [req.user?.userId]);
+    const noCart = results.rows;
+    if (!noCart) throw new ClientError(404, 'User Id not found!');
+    res.sendStatus(204);
+  } catch (error) {
+    next(error);
+  }
+});
+
 app.get('/api/catalog/cart', authMiddleware, async (req, res, next) => {
   try {
     const sql = `
@@ -224,12 +240,13 @@ app.delete(
   async (req, res, next) => {
     try {
       const { productId, size } = req.params;
+      const { userId } = req.body;
       const sql = `
        delete from "cartItems"
-       where "productId" = $1 and "size" =$2
+       where "productId" = $1 and "size" =$2 and "userId" = $3
        returning *
                  `;
-      const params = [productId, size];
+      const params = [productId, size, req.user?.userId];
       const results = await db.query(sql, params);
       const deletedCartItem = results.rows[0];
       if (!deletedCartItem)
@@ -247,15 +264,16 @@ app.put(
   async (req, res, next) => {
     try {
       const { productId } = req.params;
-      const { quantity, size } = req.body;
+      const { quantity, size, userId } = req.body;
       const sql = `
     update "cartItems"
     set "quantity" = $1
     where "productId" = $2
     and "size" = $3
+    and "userId" = $4
     returning*
     `;
-      const params = [quantity, productId, size];
+      const params = [quantity, productId, size, req.user?.userId];
       const results = await db.query(sql, params);
       const updatedProduct = results.rows[0];
       if (!updatedProduct) {
@@ -309,12 +327,14 @@ app.delete(
   async (req, res, next) => {
     try {
       const { productId } = req.params;
+      const { userId } = req.body;
       const sql = `
        delete from "wishlists"
        where "productId" = $1
+       and "userId" = $2
        returning *
                  `;
-      const params = [productId];
+      const params = [productId, req.user?.userId];
       const results = await db.query(sql, params);
       const deletedCartItem = results.rows[0];
       if (!deletedCartItem)
